splitting commands.py for simplicity
